Fix Landing Page
Resize the top gif and replace the broken "Features" button.

The top gif was set to 150px height which seemingly was ignored before
and was now honored with the hugo & syna update. The gif is now sized
appropriately.

There was no '#features' anchor on the page, which lead to the
"Features" button being broken. The button is now replaced with a
"Components" button.